requirements.txt
mysql-replication>=0.4.1
elasticsearch>=1.2.0